+++ src/main/java/domain/cards/gamercards/DealerCards.java
@@ -4,7 +4,7 @@ import domain.cards.Card;
 
 import java.util.List;
 
-public class DealerCards extends GamerCards {
+public class DealerCards extends PlayerCards {
 
     private static final int DEALER_HIT_THRESHOLD = 16;
     private static final int FIRST_CARD_INDEX = 0;
