+++ src/main/java/domain/gamer/Dealer.java
@@ -3,7 +3,7 @@ package domain.gamer;
 import domain.cards.Card;
 import domain.cards.gamercards.DealerCards;
 
-public class Dealer extends Gamer {
+public class Dealer extends Player {
 
     private static final String DEALER_NAME = "딜러";
 
@@ -12,8 +12,7 @@ public class Dealer extends Gamer {
     }
 
     public boolean canHit() {
-        DealerCards dealerCards = (DealerCards) cards;
-        return dealerCards.hasScoreUnderHitThreshold();
+        return ((DealerCards) cards).hasScoreUnderHitThreshold();
     }
 
     public Card openOneCard() {
